+++ src/main/java/com/finanzas/finanzasback/resource/FeeReceiptResource.java
@@ -11,17 +11,18 @@ public class FeeReceiptResource {
     private Date issue_date;
 
     private Date payment_date;
-    private Float     total_to_receive;
-    private Float retention;
-    private Float    total_final_costs;
-    private Float total_starting_costs;
-    private Float   discount_rate;
-    private Float discount;
-    private Float    net_worth;
-    private Float received_value;
-    private Float     delivered_value;
-    private Float tcea;
+    private double     total_to_receive;
+    private double retention;
+    private double    total_final_costs;
+    private double total_starting_costs;
+    private double   discount_rate;
+    private double discount;
+    private double    net_worth;
+    private double received_value;
+    private double     delivered_value;
+    private double tcea;
     private int dias;
+
 
     public Long getId() {
         return id;
@@ -55,83 +56,83 @@ public class FeeReceiptResource {
         this.payment_date = payment_date;
     }
 
-    public Float getTotal_to_receive() {
+    public double getTotal_to_receive() {
         return total_to_receive;
     }
 
-    public void setTotal_to_receive(Float total_to_receive) {
+    public void setTotal_to_receive(double total_to_receive) {
         this.total_to_receive = total_to_receive;
     }
 
-    public Float getRetention() {
+    public double getRetention() {
         return retention;
     }
 
-    public void setRetention(Float retention) {
+    public void setRetention(double retention) {
         this.retention = retention;
     }
 
-    public Float getTotal_final_costs() {
+    public double getTotal_final_costs() {
         return total_final_costs;
     }
 
-    public void setTotal_final_costs(Float total_final_costs) {
+    public void setTotal_final_costs(double total_final_costs) {
         this.total_final_costs = total_final_costs;
     }
 
-    public Float getTotal_starting_costs() {
+    public double getTotal_starting_costs() {
         return total_starting_costs;
     }
 
-    public void setTotal_starting_costs(Float total_starting_costs) {
+    public void setTotal_starting_costs(double total_starting_costs) {
         this.total_starting_costs = total_starting_costs;
     }
 
-    public Float getDiscount_rate() {
+    public double getDiscount_rate() {
         return discount_rate;
     }
 
-    public void setDiscount_rate(Float discount_rate) {
+    public void setDiscount_rate(double discount_rate) {
         this.discount_rate = discount_rate;
     }
 
-    public Float getDiscount() {
+    public double getDiscount() {
         return discount;
     }
 
-    public void setDiscount(Float discount) {
+    public void setDiscount(double discount) {
         this.discount = discount;
     }
 
-    public Float getNet_worth() {
+    public double getNet_worth() {
         return net_worth;
     }
 
-    public void setNet_worth(Float net_worth) {
+    public void setNet_worth(double net_worth) {
         this.net_worth = net_worth;
     }
 
-    public Float getReceived_value() {
+    public double getReceived_value() {
         return received_value;
     }
 
-    public void setReceived_value(Float received_value) {
+    public void setReceived_value(double received_value) {
         this.received_value = received_value;
     }
 
-    public Float getDelivered_value() {
+    public double getDelivered_value() {
         return delivered_value;
     }
 
-    public void setDelivered_value(Float delivered_value) {
+    public void setDelivered_value(double delivered_value) {
         this.delivered_value = delivered_value;
     }
 
-    public Float getTcea() {
+    public double getTcea() {
         return tcea;
     }
 
-    public void setTcea(Float tcea) {
+    public void setTcea(double tcea) {
         this.tcea = tcea;
     }
